refactor(services): merge lib into services
