--- src/lib/README.md
@@ -1,5 +1,0 @@
-# lib
-
-TODO: Merge this into services
-
-This directory contains narrowly-scoped utility code that is needed by Farmhand Shuffle, but is not specific to the project. This includes solutions for general needs such as standard algorithms and data structures.
