+++ src/services/README.md
@@ -1,5 +1,5 @@
 # Game Services
 
-This directory contains services that are needed by Farmhand Shuffle, but are not specific to the project. This includes solutions for general needs such as random number generation. Services provide a set of related functionality (generally in the form of a class), whereas lib functions provide narrowly-scoped, one-off utilities.
+This directory contains services that are needed by Farmhand Shuffle, but are not specific to the project. This includes solutions for general needs such as standard data structure and algorithms.
 
 The service pattern is inspired by [Angular services](https://angular.io/guide/architecture-services).
